// luckynumberapp/types/user.ts
export interface User {
  city: string | null | undefined;
  country: string | null | undefined;
  created_at: string | null | undefined;
  dob: string | null | undefined;
  fname: string | null | undefined;
  gender: string | null;
  id: string;
  is_complete: boolean;
  lname: string | null | undefined;
  street: string | null | undefined;
  zipcode: number | null | undefined;
}